HTBHF-1551 rename get previous path files
