+++ src/web/routes/application/middleware/configure-get.js
@@ -1,4 +1,4 @@
-const { getPreviousPath } = require('./get-previous-page')
+const { getPreviousPath } = require('./get-previous-path')
 
 const configureGet = (steps, step) => (req, res, next) => {
   res.locals.previous = getPreviousPath(steps, step, req.session)
