+++ src/web/routes/application/middleware/configure-post.js
@@ -1,4 +1,4 @@
-const { getPreviousPath } = require('./get-previous-page')
+const { getPreviousPath } = require('./get-previous-path')
 
 const configurePost = (steps, step) => (req, res, next) => {
   res.locals.errorTitleText = req.t('validation:errorTitleText')
